pyspark_requirements.txt
# jupyterlab-server 2.25.2 requires jsonschema>=4.18.0, but you have jsonschema 3.2.0 which is incompatible.
# jupyterlab-server 2.25.2 requires requests>=2.31, but you have requests 2.26.0 which is incompatible.
# pandas
# pandasql
# jupyter-server 2.10.1 requires jupyter-core!=5.0.*,>=4.12, but you have jupyter-core 4.10.0 which is incompatible.
# jupyter-server 2.10.1 requires send2trash>=1.8.2, but you have send2trash 1.8.0 which is incompatible.
# jupyter-server 2.10.1 requires traitlets>=5.6.0, but you have traitlets 5.3.0 which is incompatible.
jupyter-core==4.12
send2trash==1.8.2
traitlets==5.6.0
altair
matplotlib
sympy
pyodbc
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
email_validator
flask_sqlalchemy
WTForms==2.3.3
flask-wtf
email_validator
flask_sqlalchemy
flask==1.1.2
Flask-reCaptcha==0.4.2
stripe
# ecdsa
# paramiko
# pycrypto
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
coverage==6.2
cryptography==36.0.1
dnspython==2.1.0
email-validator==1.1.3
fabric==2.6.0
Flask==1.1.2
Flask-ReCaptcha==0.4.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
invoke==1.6.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.0.1
paramiko==2.8.1
pathlib2==2.3.6
pycparser==2.21
PyNaCl==1.4.0
requests==2.31
six==1.16.0
#SQLAlchemy==1.4.27
SQLAlchemy
stripe==2.63.0
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
wtforms-recaptcha==0.3.2
flask_mail
blinker
numpy
pandas
altair
pandasql
matplotlib
vega_datasets
jsonschema==4.18.0
sympy
vega==1.3
jupyter
pyodbc
vl-convert-python
papermill
xlsxwriter
openpyxl
randomname
mysql-connector-python
faker==19.6.2
fastapi[all]
sqlalchemy
databases
psycopg2-binary
asyncpg
sqlacodegen==3.0.0b2
pydantic_sqlalchemy
hide_code
jupyterlab-hide-code
redis
boto3
redshift_connector
